requirements.txt
bcrypt==5.0.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
firebase-admin==6.4.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
httplib2==0.31.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
msgpack==1.1.2
mysql-connector-python==8.2.0
numpy==1.26.3
pandas==2.1.4
proto-plus==1.26.1
protobuf==4.21.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.8.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
rsa==4.9.1
schedule==1.2.0
scikit-learn==1.4.0
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.0
Werkzeug==3.1.4
xgboost==2.0.3
zipp==3.23.0
